functions/authorizer/requirements.txt
python-jose
boto3
